feat: add lib, remove wallet
